# template/api/system/v1/menu_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v6.31.1
// source: system/v1/menu.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationMenuCheckMenuPermission = "/api.system.v1.Menu/CheckMenuPermission"
const OperationMenuCheckMultiplePermissions = "/api.system.v1.Menu/CheckMultiplePermissions"
const OperationMenuCreateMenu = "/api.system.v1.Menu/CreateMenu"
const OperationMenuDeleteMenu = "/api.system.v1.Menu/DeleteMenu"
const OperationMenuGetMenu = "/api.system.v1.Menu/GetMenu"
const OperationMenuListAllMenu = "/api.system.v1.Menu/ListAllMenu"
const OperationMenuListMenu = "/api.system.v1.Menu/ListMenu"
const OperationMenuUpdateMenu = "/api.system.v1.Menu/UpdateMenu"

type MenuHTTPServer interface {
	// CheckMenuPermission Added menu permission endpoints
	// CheckMenuPermission checks if a user has permission to perform an action on
	// a specific menu
	CheckMenuPermission(context.Context, *CheckMenuPermissionRequest) (*CheckMenuPermissionReply, error)
	// CheckMultiplePermissions CheckMultiplePermissions checks multiple menu permissions for a user in a
	// single request
	CheckMultiplePermissions(context.Context, *CheckMultiplePermissionsRequest) (*CheckMultiplePermissionsReply, error)
	// CreateMenu CreateMenu creates a new menu item in the system
	CreateMenu(context.Context, *CreateMenuRequest) (*CreateMenuReply, error)
	// DeleteMenu DeleteMenu deletes a menu item by its ID
	DeleteMenu(context.Context, *DeleteMenuRequest) (*DeleteMenuReply, error)
	// GetMenu GetMenu retrieves detailed information about a specific menu item
	GetMenu(context.Context, *GetMenuRequest) (*GetMenuReply, error)
	// ListAllMenu ListAllMenu retrieves all menu items in the system as a tree structure
	ListAllMenu(context.Context, *ListAllMenuRequest) (*ListAllMenuReply, error)
	// ListMenu ListMenu retrieves a paginated list of menu items with optional filtering
	ListMenu(context.Context, *ListMenuRequest) (*ListMenuReply, error)
	// UpdateMenu UpdateMenu updates an existing menu item's attributes
	UpdateMenu(context.Context, *UpdateMenuRequest) (*UpdateMenuReply, error)
}

func RegisterMenuHTTPServer(s *http.Server, srv MenuHTTPServer) {
	r := s.Route("/")
	r.POST("/api/v1/system/menu", _Menu_CreateMenu0_HTTP_Handler(srv))
	r.POST("/api/v1/system/menu/update", _Menu_UpdateMenu0_HTTP_Handler(srv))
	r.POST("/api/v1/system/menu/delete", _Menu_DeleteMenu0_HTTP_Handler(srv))
	r.GET("/api/v1/system/menu/get", _Menu_GetMenu0_HTTP_Handler(srv))
	r.GET("/api/v1/system/menu/list", _Menu_ListMenu0_HTTP_Handler(srv))
	r.GET("/api/v1/system/menu/all", _Menu_ListAllMenu0_HTTP_Handler(srv))
	r.POST("/api/v1/system/menu/check", _Menu_CheckMenuPermission0_HTTP_Handler(srv))
	r.POST("/api/v1/system/menu/check/multi", _Menu_CheckMultiplePermissions0_HTTP_Handler(srv))
}

func _Menu_CreateMenu0_HTTP_Handler(srv MenuHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateMenuRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMenuCreateMenu)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateMenu(ctx, req.(*CreateMenuRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateMenuReply)
		return ctx.Result(200, reply)
	}
}

func _Menu_UpdateMenu0_HTTP_Handler(srv MenuHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateMenuRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMenuUpdateMenu)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateMenu(ctx, req.(*UpdateMenuRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateMenuReply)
		return ctx.Result(200, reply)
	}
}

func _Menu_DeleteMenu0_HTTP_Handler(srv MenuHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteMenuRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMenuDeleteMenu)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteMenu(ctx, req.(*DeleteMenuRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteMenuReply)
		return ctx.Result(200, reply)
	}
}

func _Menu_GetMenu0_HTTP_Handler(srv MenuHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetMenuRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMenuGetMenu)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetMenu(ctx, req.(*GetMenuRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetMenuReply)
		return ctx.Result(200, reply)
	}
}

func _Menu_ListMenu0_HTTP_Handler(srv MenuHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListMenuRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMenuListMenu)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListMenu(ctx, req.(*ListMenuRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListMenuReply)
		return ctx.Result(200, reply)
	}
}

func _Menu_ListAllMenu0_HTTP_Handler(srv MenuHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListAllMenuRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMenuListAllMenu)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListAllMenu(ctx, req.(*ListAllMenuRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListAllMenuReply)
		return ctx.Result(200, reply)
	}
}

func _Menu_CheckMenuPermission0_HTTP_Handler(srv MenuHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CheckMenuPermissionRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMenuCheckMenuPermission)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CheckMenuPermission(ctx, req.(*CheckMenuPermissionRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CheckMenuPermissionReply)
		return ctx.Result(200, reply)
	}
}

func _Menu_CheckMultiplePermissions0_HTTP_Handler(srv MenuHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CheckMultiplePermissionsRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMenuCheckMultiplePermissions)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CheckMultiplePermissions(ctx, req.(*CheckMultiplePermissionsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CheckMultiplePermissionsReply)
		return ctx.Result(200, reply)
	}
}

type MenuHTTPClient interface {
	// CheckMenuPermission Added menu permission endpoints
	// CheckMenuPermission checks if a user has permission to perform an action on
	// a specific menu
	CheckMenuPermission(ctx context.Context, req *CheckMenuPermissionRequest, opts ...http.CallOption) (rsp *CheckMenuPermissionReply, err error)
	// CheckMultiplePermissions CheckMultiplePermissions checks multiple menu permissions for a user in a
	// single request
	CheckMultiplePermissions(ctx context.Context, req *CheckMultiplePermissionsRequest, opts ...http.CallOption) (rsp *CheckMultiplePermissionsReply, err error)
	// CreateMenu CreateMenu creates a new menu item in the system
	CreateMenu(ctx context.Context, req *CreateMenuRequest, opts ...http.CallOption) (rsp *CreateMenuReply, err error)
	// DeleteMenu DeleteMenu deletes a menu item by its ID
	DeleteMenu(ctx context.Context, req *DeleteMenuRequest, opts ...http.CallOption) (rsp *DeleteMenuReply, err error)
	// GetMenu GetMenu retrieves detailed information about a specific menu item
	GetMenu(ctx context.Context, req *GetMenuRequest, opts ...http.CallOption) (rsp *GetMenuReply, err error)
	// ListAllMenu ListAllMenu retrieves all menu items in the system as a tree structure
	ListAllMenu(ctx context.Context, req *ListAllMenuRequest, opts ...http.CallOption) (rsp *ListAllMenuReply, err error)
	// ListMenu ListMenu retrieves a paginated list of menu items with optional filtering
	ListMenu(ctx context.Context, req *ListMenuRequest, opts ...http.CallOption) (rsp *ListMenuReply, err error)
	// UpdateMenu UpdateMenu updates an existing menu item's attributes
	UpdateMenu(ctx context.Context, req *UpdateMenuRequest, opts ...http.CallOption) (rsp *UpdateMenuReply, err error)
}

type MenuHTTPClientImpl struct {
	cc *http.Client
}

func NewMenuHTTPClient(client *http.Client) MenuHTTPClient {
	return &MenuHTTPClientImpl{client}
}

// CheckMenuPermission Added menu permission endpoints
// CheckMenuPermission checks if a user has permission to perform an action on
// a specific menu
func (c *MenuHTTPClientImpl) CheckMenuPermission(ctx context.Context, in *CheckMenuPermissionRequest, opts ...http.CallOption) (*CheckMenuPermissionReply, error) {
	var out CheckMenuPermissionReply
	pattern := "/api/v1/system/menu/check"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMenuCheckMenuPermission))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// CheckMultiplePermissions CheckMultiplePermissions checks multiple menu permissions for a user in a
// single request
func (c *MenuHTTPClientImpl) CheckMultiplePermissions(ctx context.Context, in *CheckMultiplePermissionsRequest, opts ...http.CallOption) (*CheckMultiplePermissionsReply, error) {
	var out CheckMultiplePermissionsReply
	pattern := "/api/v1/system/menu/check/multi"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMenuCheckMultiplePermissions))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// CreateMenu CreateMenu creates a new menu item in the system
func (c *MenuHTTPClientImpl) CreateMenu(ctx context.Context, in *CreateMenuRequest, opts ...http.CallOption) (*CreateMenuReply, error) {
	var out CreateMenuReply
	pattern := "/api/v1/system/menu"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMenuCreateMenu))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// DeleteMenu DeleteMenu deletes a menu item by its ID
func (c *MenuHTTPClientImpl) DeleteMenu(ctx context.Context, in *DeleteMenuRequest, opts ...http.CallOption) (*DeleteMenuReply, error) {
	var out DeleteMenuReply
	pattern := "/api/v1/system/menu/delete"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMenuDeleteMenu))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetMenu GetMenu retrieves detailed information about a specific menu item
func (c *MenuHTTPClientImpl) GetMenu(ctx context.Context, in *GetMenuRequest, opts ...http.CallOption) (*GetMenuReply, error) {
	var out GetMenuReply
	pattern := "/api/v1/system/menu/get"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMenuGetMenu))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListAllMenu ListAllMenu retrieves all menu items in the system as a tree structure
func (c *MenuHTTPClientImpl) ListAllMenu(ctx context.Context, in *ListAllMenuRequest, opts ...http.CallOption) (*ListAllMenuReply, error) {
	var out ListAllMenuReply
	pattern := "/api/v1/system/menu/all"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMenuListAllMenu))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListMenu ListMenu retrieves a paginated list of menu items with optional filtering
func (c *MenuHTTPClientImpl) ListMenu(ctx context.Context, in *ListMenuRequest, opts ...http.CallOption) (*ListMenuReply, error) {
	var out ListMenuReply
	pattern := "/api/v1/system/menu/list"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMenuListMenu))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateMenu UpdateMenu updates an existing menu item's attributes
func (c *MenuHTTPClientImpl) UpdateMenu(ctx context.Context, in *UpdateMenuRequest, opts ...http.CallOption) (*UpdateMenuReply, error) {
	var out UpdateMenuReply
	pattern := "/api/v1/system/menu/update"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMenuUpdateMenu))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}