# template/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/system/menu:
        post:
            tags:
                - Menu
            description: CreateMenu creates a new menu item in the system
            operationId: Menu_CreateMenu
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.CreateMenuRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.CreateMenuReply'
...
    /api/v1/system/menu/check:
        post:
            tags:
                - Menu
            description: |-
                Added menu permission endpoints
                 CheckMenuPermission checks if a user has permission to perform an action on
                 a specific menu
            operationId: Menu_CheckMenuPermission
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.CheckMenuPermissionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.CheckMenuPermissionReply'
    /api/v1/system/menu/check/multi:
        post:
            tags:
                - Menu
            description: |-
                CheckMultiplePermissions checks multiple menu permissions for a user in a
                 single request
            operationId: Menu_CheckMultiplePermissions
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.CheckMultiplePermissionsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.CheckMultiplePermissionsReply'
    /api/v1/system/menu/delete:
        post:
            tags:
                - Menu
            description: DeleteMenu deletes a menu item by its ID
            operationId: Menu_DeleteMenu
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.DeleteMenuRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.DeleteMenuReply'
    /api/v1/system/menu/get:
        get:
            tags:
                - Menu
            description: GetMenu retrieves detailed information about a specific menu item
            operationId: Menu_GetMenu
            parameters:
                - name: id
                  in: query
                  description: 菜单ID
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.GetMenuReply'
    /api/v1/system/menu/list:
        get:
            tags:
                - Menu
            description: ListMenu retrieves a paginated list of menu items with optional filtering
            operationId: Menu_ListMenu
            parameters:
                - name: currentPage
                  in: query
                  description: 当前页
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  description: 每页条数
                  schema:
                    type: integer
                    format: int32
                - name: name
                  in: query
                  description: 菜单名称，用于筛选
                  schema:
                    type: string
                - name: isShow
                  in: query
                  description: 状态筛选 0 隐藏 1 显示
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.ListMenuReply'
    /api/v1/system/menu/update:
        post:
            tags:
                - Menu
            description: UpdateMenu updates an existing menu item's attributes
            operationId: Menu_UpdateMenu
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.UpdateMenuRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.UpdateMenuReply'
...
    /api/v1/system/role/menu:
        post:
            tags:
                - Role
            description: UpdateRoleMenus updates the menu permissions for a specific role
            operationId: Role_UpdateRoleMenus
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.UpdateRoleMenusRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.UpdateRoleMenusReply'
    /api/v1/system/role/menus:
        get:
            tags:
                - Role
            description: ListRoleMenus retrieves the list of menu IDs assigned to a specific role
            operationId: Role_ListRoleMenus
            parameters:
                - name: roleId
                  in: query
                  description: 角色ID
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.ListRoleMenusReply'
...
components:
    schemas:
...
        api.system.v1.CheckMenuPermissionReply:
            type: object
            properties:
                hasPermission:
                    type: boolean
                    description: 是否有权限
        api.system.v1.CheckMenuPermissionRequest:
            type: object
            properties:
                userId:
                    type: string
                    description: 用户ID
                menuId:
                    type: string
                    description: 菜单ID
                action:
                    type: string
                    description: 操作类型
            description: Added menu permission messages
        api.system.v1.CheckMultiplePermissionsReply:
            type: object
            properties:
                results:
                    type: object
                    additionalProperties:
                        type: boolean
                    description: 权限检查结果，key为菜单ID，value为是否有权限
        api.system.v1.CheckMultiplePermissionsRequest:
            type: object
            properties:
                userId:
                    type: string
                    description: 用户ID
                permissions:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.system.v1.MenuPermission'
                    description: 权限列表
...
        api.system.v1.CreateMenuReply:
            type: object
            properties: {}
        api.system.v1.CreateMenuRequest:
            type: object
            properties:
                pid:
                    type: string
                    description: 父ID
                name:
                    type: string
                    description: 规则名称
                title:
                    type: string
                    description: 规则名称
                icon:
                    type: string
                    description: 图标
                condition:
                    type: string
                    description: 条件
                remark:
                    type: string
                    description: 备注
                menuType:
                    type: integer
                    description: 类型 0目录 1菜单 2按钮
                    format: int32
                weight:
                    type: integer
                    description: 权重
                    format: int32
                isShow:
                    type: integer
                    description: 显示状态
                    format: int32
                path:
                    type: string
                    description: 路由地址
                component:
                    type: string
                    description: 组件路径
                isLink:
                    type: integer
                    description: 是否外链 1是 0否
                    format: int32
                moduleType:
                    type: string
                    description: 所属模块
                modelId:
                    type: integer
                    description: 模型ID
                    format: int32
                isIframe:
                    type: integer
                    description: 是否内嵌iframe
                    format: int32
                isCached:
                    type: integer
                    description: 是否缓存
                    format: int32
                redirect:
                    type: string
                    description: 路由重定向地址
                isAffix:
                    type: integer
                    description: 是否固定
                    format: int32
                linkUrl:
                    type: string
                    description: 链接地址
...
        api.system.v1.DeleteMenuReply:
            type: object
            properties: {}
        api.system.v1.DeleteMenuRequest:
            type: object
            properties:
                id:
                    type: string
                    description: 菜单ID
...
        api.system.v1.GetMenuReply:
            type: object
            properties:
                menu:
                    $ref: '#/components/schemas/api.system.v1.MenuInfo'
...
        api.system.v1.ListMenuReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.system.v1.MenuInfo'
                    description: 菜单列表（树形结构）
                total:
                    type: string
                    description: 总条数
...
        api.system.v1.ListRoleMenusReply:
            type: object
            properties:
                menuIds:
                    type: array
                    items:
                        type: string
                    description: 菜单ID
...
        api.system.v1.MenuInfo:
            type: object
            properties:
                id:
                    type: string
                    description: ID
                pid:
                    type: string
                    description: 父ID
                name:
                    type: string
                    description: 规则名称
                title:
                    type: string
                    description: 规则名称
                icon:
                    type: string
                    description: 图标
                condition:
                    type: string
                    description: 条件
                remark:
                    type: string
                    description: 备注
                menuType:
                    type: integer
                    description: 类型 0目录 1菜单 2按钮
                    format: int32
                weight:
                    type: integer
                    description: 权重
                    format: int32
                isShow:
                    type: integer
                    description: 显示状态 0 隐藏 1 显示
                    format: int32
                path:
                    type: string
                    description: 路由地址
                component:
                    type: string
                    description: 组件路径
                isLink:
                    type: integer
                    description: 是否外链 1是 0否
                    format: int32
                moduleType:
                    type: string
                    description: 所属模块
                modelId:
                    type: integer
                    description: 模型ID
                    format: int32
                isIframe:
                    type: integer
                    description: 是否内嵌iframe
                    format: int32
                isCached:
                    type: integer
                    description: 是否缓存
                    format: int32
                redirect:
                    type: string
                    description: 路由重定向地址
                isAffix:
                    type: integer
                    description: 是否固定
                    format: int32
                linkUrl:
                    type: string
                    description: 链接地址
                createdBy:
                    type: string
                    description: 创建者
                updatedBy:
                    type: string
                    description: 更新者
                createdAt:
                    type: string
                    description: 创建时间
                updatedAt:
                    type: string
                    description: 更新时间
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.system.v1.MenuInfo'
                    description: 子菜单
        api.system.v1.MenuPermission:
            type: object
            properties:
                menuId:
                    type: string
                    description: 菜单ID
                action:
                    type: string
                    description: 操作类型
...
        api.system.v1.UpdateMenuReply:
            type: object
            properties: {}
        api.system.v1.UpdateMenuRequest:
            type: object
            properties:
                id:
                    type: string
                    description: 菜单ID
                name:
                    type: string
                    description: 规则名称
                title:
                    type: string
                    description: 规则名称
                icon:
                    type: string
                    description: 图标
                condition:
                    type: string
                    description: 条件
                remark:
                    type: string
                    description: 备注
                menuType:
                    type: integer
                    description: 类型 0目录 1菜单 2按钮
                    format: int32
                weight:
                    type: integer
                    description: 权重
                    format: int32
                isShow:
                    type: integer
                    description: 显示状态 0 隐藏 1 显示
                    format: int32
                path:
                    type: string
                    description: 路由地址
                component:
                    type: string
                    description: 组件路径
                isLink:
                    type: integer
                    description: 是否外链 1是 0否
                    format: int32
                moduleType:
                    type: string
                    description: 所属模块
                modelId:
                    type: integer
                    description: 模型ID
                    format: int32
                isIframe:
                    type: integer
                    description: 是否内嵌iframe
                    format: int32
                isCached:
                    type: integer
                    description: 是否缓存
                    format: int32
                redirect:
                    type: string
                    description: 路由重定向地址
                isAffix:
                    type: integer
                    description: 是否固定
                    format: int32
                linkUrl:
                    type: string
                    description: 链接地址
...
        api.system.v1.UpdateRoleMenusReply:
            type: object
            properties: {}
        api.system.v1.UpdateRoleMenusRequest:
            type: object
            properties:
                roleId:
                    type: string
                menuIds:
                    type: array
                    items:
                        type: string
...